requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
docstring_parser==0.17.0
google==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
llm-axe==1.1.9
numpy==2.3.2
ollama==0.5.3
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.9.0
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
threadpoolctl==3.6.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
